Refactor and Fix issue code per review
